Fixed bug in generating weather features.
The bug was in the mesowest api caller. The code should fill in the 'peak_wind_speed' of a given station with the maximum wind_gust for a given day. Previously, there was an error in how it assigned the 'peak_wind_speed' where it would throw an error if there was no 'peak_wind_speed' value in the api call. It works as it should now, and should also correct errors in the 'peak_wind_direction'.